requirements.txt
keras==2.15.0
matplotlib==3.7.0
numpy==1.23.2
opencv_python==4.7.0.72
patchify==0.2.3
Pillow==10.0.0
streamlit==1.18.1
tensorflow==2.15.0